requirements.txt

matplotlib==2.1.0
numpy==1.14.2
Fiona==1.7.10
geojson==2.3.0
Shapely==1.6.2
descartes==1.1.0
psycopg2==2.7.4
scipy==1.0.1
osr==0.0.1
pyshp==1.2.12
scikit_learn==0.19.1
scikit-image
opencv-python==3.3.1.11
tensorflow-gpu
Pillow
tqdm
